Added float tensor casts to datasets

+++ datasets.py
@@ -46,7 +46,7 @@ class MFNetDataset(Dataset):
 
         if self.label_map is not None:
             label = np.array(self.label_map)[label]
-        return image, label.long(), name
+        return image.float(), label.long(), name
 
     def get_test_item(self, index):
         name = self.names[index]
@@ -138,7 +138,7 @@ class HeatNetDataset(Dataset):
                 image, label = func(image, label)
             if self.label_map is not None:
                 label = np.array(self.label_map)[label]
-            return image, label.long()
+            return image.float(), label.long()
         else:
             image = self.get_train_item(*self.names[index])
             for func in self.transform:
